moves auth folder ouside of components
