+++ client/src/auth/PrivateRoute.jsx
@@ -1,5 +1,5 @@
-import React, { useEffect } from './node_modules/react';
-import { Route } from './node_modules/react-router-dom';
+import React, { useEffect } from 'react';
+import { Route } from 'react-router-dom';
 import { useAuth0 } from './authWrapper';
 
 const PrivateRoute = ({ component: Component, path, ...rest }) => {
