+++ client/src/auth/authWrapper.jsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect, useContext } from 'react';
 import createAuth0Client from '@auth0/auth0-spa-js';
 import { useStateValue } from 'react-conflux';
-import { globalContext } from '../../store/reducers/globalReducer';
+import { globalContext } from '../store/reducers/globalReducer';
 import { register } from './authQueries';
 
 const DEFAULT_REDIRECT_CALLBACK = () =>
